refactor(actions): move action-models into 'type' package
